python-scripts/requirements.txt
docaligner-docsaid>=1.1.0
capybara-docsaid>=0.5.1
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
